requirements.txt
torch
ftfy
regex
tqdm
Pillow
numpy
PyMuPDF
git+https://github.com/openai/CLIP.git

